+++ crates/dioxus-bootstrap/src/lib.rs
@@ -72,6 +72,7 @@ pub mod grid;
 pub mod head;
 pub mod icon;
 pub mod input_group;
+mod keyboard;
 pub mod list_group;
 pub mod modal;
 pub mod nav;